Revert "Call SetName only if the transaction is doing writes"
Summary:
This reverts an earlier optimization to only call `SetName` for transactions that do writes, by instrumenting all the places where writes were happening to call `SetName` on demand. The concern is that we might be missing places, so it was just safer to set the name at the start of the transaction.

This revert adds 2 additional things:
1. Calling `SetName` at the beginning only happens for write unprepared, so that we limit the read query regression to that write policy. For other policies, we still call `SetName` at prepare time.
2. Some internal operations also start transactions but without a valid xid (this was exposed after master-info-repository was set to TABLE). These don't go through 2pc though, so it's okay to skip setting name.

Reviewed By: hermanlee

Differential Revision: D22801035

fbshipit-source-id: 2ff451c0f3a