Remove fast unpack code path in bypass
Summary: The fast unpacking code path in bypass has intrinsic understanding of MyRocks storage format and was mostly taken from Yoshi's prototype. Given that MyRocks format will keep evolving, and we'll be also adding new features, having a custom decoding path is risky and hard to maintain. In my testing it also made relatively little difference in the latest benchmark that I have, so this just takes it out

Reviewed By: hermanlee

Differential Revision: D22837401

fbshipit-source-id: 44b534e82e6