PS-7027: MyRocks: Add SK index-only scans for all remaining collations
https://jira.percona.com/browse/PS-7027

1. Index-only scans with SK are supported by decoding the memcomparable form back into the original form.
Currently we support SK index-only scans for:
- `binary`, `utf8_bin`, and `utf8mb4_bin` collations
- collations based on `my_collation_8bit_simple_ci_handler` (e.g. `latin1_swedish_ci`, `cp850_general_ci`, `latin1_german1_ci`, `cp1251_ukrainian_ci`, `cp866_general_ci`)
- collations based on `my_collation_8bit_bin_handler` (i.e. `latinX_bin`, `cp12XX_bin`, `cp85X_bin`, `armscii8_bin`, `ascii_bin` and many more)

This patch turns on support for all remaining collations by storing entire copy of original field in the value.
The downside is that extra storage is used for the index.

2. Add backward compatibility with PRIMARY_FORMAT_VERSION_COLL and SECONDARY_FORMAT_VERSION_COLL
after introducing a support for index-only scans with SK for all previously-unsupported collations
by storing entire copy of original field in `unpack_info`.