Limit memory usage used by quick ranges during range analysis
Summary:
During range optimization, a tree of SEL_ARG objects are built, representing the set of intervals that we need to examine for the query. A description of the tree representation exists here (https://github.com/facebook/mysql-5.6/blob/c2a7d7274b0d9c22a8e7affc41fb53f6ef7083f3/sql/opt_range.cc#L188).

To prevent the tree from consuming too much memory (eg. from very large IN-lists), a variable called range_optimizer_max_mem_size exists to limit memory usage.

However, once range analysis is finished, the tree is then flattened into a list of ranges, and this can grow large very quickly. For example, a query like `SELECT .. WHERE a in (1, 2, ... 1000) and b in (1, 2, ... 1000)` will contain 1000000 ranges. There is no memory limit for these allocations, and this can cause the server to OOM.

To address this, we will also apply the same range_optimizer_max_mem_size limit on the list of ranges. We could have also had a separate limit for quick ranges, but we would need the error handler to be aware of which MEM_ROOT object was used. Also, it seemed simpler for the end user to not have to distinguish between the cases, since the tree gets freed after building the ranges anyway.

This change also includes a feature to optionally error out the query when memory limits are reached. The problem today is that when range analysis runs out of memory, we fall back to full index scans. It may be better just to error out.

Some implementation notes:
- QUICK_* objects (eg. QUICK_RANGE_SELECT) contain the QUICK_RANGEs needed for execution. Each of these QUICK objects contain a MEM_ROOT object already. Presumably the intention is that all objects belonging to the QUICK object should be backed by its MEM_ROOT, and this is what this fix does. We are making sure that all allocations for QUICK_RANGEs belonging to a QUICK object come from its associated MEM_ROOT.
- Some QUICK objects contain other children QUICK objects (eg. QUICK_GROUP_MIN_MAX_SELECT may contain a QUICK_RANGE_SELECT object for some bookkeeping). In these cases, the children QUICK objects do not use their own MEM_ROOTs (and it is zero'd out) and will use their parent's MEM_ROOT instead. Therefore, we should always be using `parent_alloc` for allocations if possible.
- A limit checking MEM_ROOT will still allocate memory even when it has reached its limit. This means that the caller is responsible for checking for errors, and for ending safely whatever recursion/loop creating these allocations.

Filed an upstream bug:
https://bugs.mysql.com/100595

Reviewed By: hermanlee

Differential Revision: D23191795

fbshipit-source-id: 1b5fbc92508