For raft relay logs adding hlc value of 0
Summary:
In Raft when HLC is enabled, we want both binlogs and relay
logs which are both raft logs should have the same size and same binary
match from the end of the previous HLC event, or start of the
transactions. This ensures that the end log pos is consistent. This is
required so that the relay-log.info file has the proper position of
applied events. In order to achieve this, we have to make sure that the
top part of the binlog file is identical (since this is not replicated).
The top part contains format descriptor, previous gtid and previous hlc
events. In this diff, we make sure that a previous hlc is also put in
raft relay logs.

Reviewed By: bhatvinay

Differential Revision: D23634741

fbshipit-source-id: 5932f0fbad5