Switch back to MySQL-python for CentOS 8.0 and convert to python3
Summary:
Given that we have recently added MySQLdb support in CentOS 8.0 for python3, we can now switch back to MySQLdb but that also means we need to convert to python3. We have two python3 in the system and we need to call /usr/bin/python3 (which is the one that has MySQLdb) instead of python3 (which symlinks to /usr/local/bin/python3)

In more details:
* Revert changes that uses pymysql
* Change python2 -> /usr/bin/python3
* Change print to a function call, and change >> log to file=log
* import stringIO is removed
* Convert MySQLdb returned string to bytes for hashing (it seems that MySQLdb in different OS has slightly different behavior and needs follow up)
* catch exception, e to catch Exception as e
* xrange -> range
* added `from MySQLdb.constants import *` to import constants
* string.letters -> string.ascii_letters

NOTE: Not every single python code has been converted to 3.0. Only those have MySQLdb dependency has been converted. Converting everything to Python 3.0 is a much larger undertaking that probably should go to Oracle.

Reviewed By: atish2196

Differential Revision: D23684770

fbshipit-source-id: e9122839973