Remove checking for connect stage after mysql_close
Summary:
The main.mysql_client_test test has been failing because the test tries to access the MYSQL object after it has been freed via `mysql_close`. Fix by removing this part of the test.

It is possible for `mysql_close` to not have freed the MYSQL object depending on how `mysql_init` was called, but in our usage of `mysql_init`, the object would always be freed.

Squash with: D23387197

Reviewed By: aditya-jalan, jkedgar

Differential Revision: D23796027

fbshipit-source-id: 6c8e1031386