Change HA_ERR_ROCKSDB_FIRST to 500 to avoid HA_ERR_LAST change causing test failures
Summary: There is a recent change that updates HA_ERR_LAST broke a few MyRocks test that rely on error code because it shifts the MyRocks error code that are based on HA_ERR_LAST. In MyRocks 8.0 we had the same problem and simply updated error code to start from 500 instead. This is a port of that change.

Reviewed By: hermanlee

Differential Revision: D23924605

fbshipit-source-id: 975b7c33f63