Use $ORIGIN to load librocks_object_factory.so alongside mysqld
Summary: In wsenv, mysqld would have a direct dependency on librocks_object_factory.so. Under development scenario, linux would always find librocks_object_factory.so in the original location where we compiled against, but at runtime in production in RPM we need a way to let linux loader find it. The correct way to do this is to use runtime search path (rpath) and set it to $ORIGIN to signal linux loader to find it alongside mysqld.

Differential Revision: D23947241