Fix admission control counters
Summary:
We have seen problems where admission control exit does not decrement running count from the correct entity, leading to underflow. This changes AC so that on enter, we record which `Ac_info` object we've entered, so that we decrement the correct object on exit.

This also fixes a race with `DROP DATABASE`. It is possible for `DROP DATABASE` to free `Ac_info`, a second query to recreate it, and a third query exiting AC to decrement the newly created `Ac_info` object, causing underflow. Instead, queries admitted in AC will hold a reference count to `Ac_info` so that on exit, they will always decrement from the same object as they entered with.

This also fixes an unrelated bug relating to the dequeue process. It seems like dequeue always happens if there are waiting queries. However, this is not necessarily desirable if the max running queries was recently lowered.

Reviewed By: george-reynya

Differential Revision: D23995710

fbshipit-source-id: 523476367db