Use extended fatal signal handler
Summary: In some cases gdb has trouble fully restoring the signal context for mysql core dump. To help see the register values in the debugger, this change installs an extended signal handler that receives the signal context and saves it on the stack to make sure it is not optimized out.

Reviewed By: yizhang82

Differential Revision: D24096305

fbshipit-source-id: cfd2a35905b