Subtract pre-existing usage from sql_stats.disk_usage columns
Summary:
Fix a bug which causes the current disk usage to be included in the statement disk usage. Typical manifestations of the bug:
- `drop temp table` usage shows the temp table size instead of 0.
- In the presence of a temp table all other temp table operations will include the existing temp table size.

Filesort shouldn't be affected but is included in the fix for consistency.

Squash with D22935862 (https://github.com/facebook/mysql-5.6/commit/a3fcc946fe58942a0de906265a56f785fdffd923).

Reviewed By: sagar0

Differential Revision: D24146869

fbshipit-source-id: 871d43d4ba3