Enable val_opid setting for prev_opid so that followers don't use 0,0
Summary:
The existing RPM for Raft already has prev_opid running with
LEADERs reporting proper prev_opid and FOLLOWERs reporting 0,0. There
was an impedance mismatch on when the plugin diff landed and when the
server diff landed, hence the followers had temporarily not been setting
it. Now we are enabling it to complete this featue.

Reviewed By: yashtc

Differential Revision: D24170387

fbshipit-source-id: 70be4e5eba3