Introduce disk_usage_period_peak vars with auto reset
Summary:
Metrics collector records disk_usage variables once every 60 sec however those infrequent samples often miss the actual usage. To mitigate this, another variable is introduced to record the peak for a period between samples. At the sample time the peak value is returned, and the variable is atomically reset to the current usage. The new session variable `reset_period_status_vars` should be set to ON in order to indicate that the `SHOW STATUS` command should reset those period variables.

This change also fixes `FLUSH STATUS` to reset the session peak to the current value instead of 0.

Reviewed By: sagar0

Differential Revision: D24280556

fbshipit-source-id: 78af613d953