Refactor ac_info switch
Summary: In preparation for using `thd->ac_node->ac_info` for connection accounting, this change refactors most of the logic of `set_session_db_helper()` into the class `Ac_switch_guard`. The main motivation is that connection accounting first obtains connection for the new db, and if successful, releases connection for the old db. So the existing `ac_info` needs to be saved, and either released at the end if the db switch is successful, or put back info `ac_node` and the new one released if switch failed.

Reviewed By: lth

Differential Revision: D24319962

fbshipit-source-id: 80dca3a0b71