Add --new-tests option
Summary:
Added a new option '--new-tests' to myqltest.sh in tools repos which is
processed by mysql-test/mysql-test-run.pl.

The option '--new-tests' runs all the tests that are either added or modified
as part of the current commit. It detects all the files that end with file
extension '.test' and which are either committed or staged or unstaged.

Reviewed By: yizhang82

Differential Revision: D24344478

fbshipit-source-id: 4629af1610c