Fix issue killing slave_stats_daemon thread when primary is stopped using SIGSTOP
Summary:
This diff fixes the bug that caused S213868

## What was the issue?

We did not configure a read and connection timeout for the connection from secondary to primary that is used to continuously send lag statistics from secondaries to primaries as implemented in the diff D21440060 (https://github.com/facebook/mysql-5.6/commit/06e73673b562438e99f83f43a3fba069d5a7e4d8)

When a primary was stopped using SIGSTOP signal and secondary tried to send lag stats to primary using this connection, this would block the `slave_stats_daemon` thread in a waiting state trying to read the response from primary that may never respond.

bart2 tried to stop the secondary using `STOP SLAVE` in this case and that led to inconsistent replica states. This was because the slave wasn't able to stop itself as one of the dependency for it to stop is to be able to stop the `slave_stats_daemon` thread which is now stuck in a never ending wait. Hence `STOP SLAVE` command would never finish.

## fix

Added an upper limit timeout of 10s for both reads and connect commands. If not successful, the thread will retry again in the next period.

Reviewed By: mzait

Differential Revision: D24536559

fbshipit-source-id: 086cf4311df