Properly adding raft term and index to apply side binlog metadata event
Summary:
When a Metadata event with Raft OpId is picked up from relay log and
applied, ev->apply_event in rpl_slave.cc stashes the raft term
and index from the event into the THD. Previous to this patch, the
stashed raft term and index was only used in before_commit hook to
determine commit and to update apply markers, but were never written
to apply side binlog. This prevented us to read binlog files of non-raft
instances tailing raft rings (via com_binlog_dump_gtid) to find their
binlog bootstrap point in the plugin.

Here we pick it up the raft opid from the THD to pass the Raft term and index
through to the metadata event of binlog/apply side. OpIds can only be
present in raft rings, hence the exposure of this code is to instances
which are tailing raft rings or raft members which are now passing OpId
to apply log as well. In all other cases Raft term and index is expected to be -1,-1
Events such as rotate, non-ops and config change rotates are still
filtered from apply side. Therefore OpId ranges can have holes and that
is OK for bootstrap, since bootstrap uses GTIDs to find opid ranges that
have been commmitted

Reviewed By: bhatvinay

Differential Revision: D24604845

fbshipit-source-id: 5f665e424ce