Throttle write traffic based on query attributes - TAO use case
Summary:
TAO team is excited to start testing with replication lag auto throttling feature but they'd like to have more control on what traffic can be throttled and what cannot, at least in the early phases.

In this diff, I have added logic for a new client attribute `mt_throttle_okay`. If this attribute is sent along with the query, we allow the query to be throttled(error/warning) else we do not throttle it.

This feature is controlled by a new boolean session variable `write_throttle_tag_only`. If set to true, we make throttling decision based on presence of `mt_throttle_okay` tag. Default value = `false`.

Reviewed By: mzait

Differential Revision: D24603291

fbshipit-source-id: 2ff6ebca1bf