Fix query tag based throttling logic for TAO use case
Summary:
with previous logic, if we want to throw error for tagged TAO queries, we also needed to set write_control_level to error. But if we do that, auto throttling logic will also start throwing errors for non TAO queries which we don't want.

In this diff, i am extending the definition of mt_throttle_okay tag. It case be set to ERROR/WARN. The value of this query attribute will override the write_control_level if query tag based throttling is turned on.

Reviewed By: mzait

Differential Revision: D24652435

fbshipit-source-id: bc7f275c033