Fix MRR to not double count ranges
Summary:
This change fixes a bug in the MyRocks implementation of MRR where we were double count ranges. During an MRR query,  `ha_rocksdb::multi_range_read_info_const` and `handler::multi_range_read_info_const` are called, and both will iterate through all ranges via `seq->init()` and `seq->next()`. However, the seq object is actually keeping count of the number of ranges during iteration, so by iterating through the ranges twice, the optimizer is actually double counting the number of ranges.

Fix this by resetting the range count in the init function.

The consequence of this bug is that we may make the wrong choices in the optimizer.

Reviewed By: yoshinorim

Differential Revision: D25103862

fbshipit-source-id: 19f3178e4c8