Fix MRR to not double count ranges
Summary:
This change fixes a bug in the MyRocks implementation of MRR where we were double count ranges. During an MRR query,  `ha_rocksdb::multi_range_read_info_const` and `handler::multi_range_read_info_const` are called, and both will iterate through all ranges via `seq->init()` and `seq->next()`. However, the seq object is actually keeping count of the number of ranges during iteration, so by iterating through the ranges twice, the optimizer is actually double counting the number of ranges.

Fix this by resetting the range count in the init function.

Porting Notes:
In 5.6, the optimizer ends up making the wrong choices due to the bad count, but in 8.0, the optimizer ends up correcting itself later and makes the right choice in the end. It's not clear if there are other edge cases that the incorrect count breaks, so we should fix it anyway in 8.0. I did not port the the mtr test because I'm not aware of these other edge cases.

Reviewed By: luqun

Differential Revision: D25105134

fbshipit-source-id: 8bffcb945b6