Support for aborting long running queries
Summary:
This diff ports the 5.6.35 diffs D23516252 (https://github.com/facebook/mysql-5.6/commit/80d7097fceef48d2fef4b8cb2dfb435865fd8470) and D23845660 (https://github.com/facebook/mysql-5.6/commit/eb2ff32db84e69596429467dda69786115ebe8f3) to 8.0.

This diff introduces the ability to abort long running statements/transactions that can potentially cause replication lag. The high level algorithm is as follows:
1. Start CPU time clock when the first row that undergoes DML is processed (inserted or deleted or updated).
2. Then, periodically after every X number of rows are processed for DML (controlled by the global variable 'write_time_check_batch'), CPU time elapsed is checked.
3. If the time elapsed is more than a threshold (controlled by global variable 'write_cpu_limit_milliseconds') then the statement/transaction is aborted.

The last step in the algorithm is rolled out to production in three stages and is controlled by global variable 'write_control_level':
* Stage 1: write_control_level='NOTE' - The statement/transaction is not aborted but a warning raised in the form a note. It is not visible to user but is mainly for internal analysis.
* Stage 2: write_control_level='WARN' - The statement/transaction is not aborted but a warning raised and is visible to user.
* Stage 3: write_control_level='ERROR' - The statement/transaction is aborted.

If the statement is aborted when write_control_level='ERROR', an entry is made into the performance schema table: write_throttling_log. The transaction_type of the entry is 'LONG', type is 'SQL_ID' and the value is the digest value (as opposed to SQL_ID in 5.6). The default value of the variable is OFF.

The diff introduced two new global variables:
- write_cpu_limit_milliseconds: Determines the CPU execution time limit allowed for DML statements/transactions expressed in milliseconds. If the CPU time spent on DML is more than the value of this variable then the statement/transaction is aborted. Default value is 0 which disables the write abort feature.
- write_time_check_batch: Determines the batch size for periodically checking if the CPU execution time for DML is more than the limit (enforced by write_cpu_limit_milliseconds). This is expressed in number of rows. Default value is 0 which disables the write abort feature.

The only major difference compared to 5.6 diff is: we have a different way of measuring the rows_deleted, rows_inserted and rows_deleted. In 5.6 these are already tracked as part of sql_stats and additional logic is added track them at transaction level. These changes are not available in 8.0 but there is some logic that tracks these stats in PSI interface. So this diff usese these to track these stats at the transaction level. These changes are mainly present in sql_class.h/.cc, handler.cc and sql_parse.cc.

Reviewed By: mzait, lth, Rahul-Singhal, george-reynya

Differential Revision: D25192588

fbshipit-source-id: 9e9334993f4